# back/docs/swagger.yaml
definitions:
...
  models.AlumneResponse:
    properties:
      ID:
        type: integer
      Nom:
        type: string
      Reserves:
        items:
          $ref: '#/definitions/models.ReservaResponse'
        type: array
      ResultatsRutinaActual:
        items:
          $ref: '#/definitions/models.UsuariResultatExerciciResponse'
        type: array
      RutinaActual:
        type: integer
      TipusUsuari:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.ExerciciInput:
    properties:
      nom:
        description: |-
          Nom es el nombre del exercici.
          required: true
          example: Extensión de tríceps en polea
        type: string
    type: object
  models.ExerciciResponse:
    properties:
      id:
        type: integer
      nom:
        type: string
    type: object
...
  models.ReservaResponse:
    properties:
      Confirmada:
        type: boolean
      Hora:
        type: string
      ID:
        type: integer
      UsuariID:
        type: integer
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
    type: object
...
  models.UsuariResultatExerciciResponse:
    properties:
      Dia:
        type: string
      ExerciciRutinaID:
        type: integer
      ID:
        type: integer
      Pes:
        type: number
      Repeticions:
        type: integer
      Series:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/entrenador/alumnes:
    get:
      consumes:
      - application/json
      description: Retrieves all the alumnes of a trainer from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AlumneResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all alumnes from  trainer
      tags:
      - Entrenador
...
  /api/exercicis:
    get:
      consumes:
      - application/json
      description: Retrieves all the exercicis from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ExerciciResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all exercicis
      tags:
      - Exercicis
    post:
      consumes:
      - application/json
      description: Creates a new exercici in the database.
      parameters:
      - description: Exercici to create
        in: body
        name: exercici
        required: true
        schema:
          $ref: '#/definitions/models.ExerciciInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ExerciciResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create an exercici
      tags:
      - Exercicis
  /api/exercicis/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an exercici from the database.
      parameters:
      - description: ID of the exercici to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete an exercici
      tags:
      - Exercicis
    get:
      consumes:
      - application/json
      description: Retrieves an exercici from the database by its ID.
      parameters:
      - description: ID of the exercici to retrieve
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ExerciciResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get an exercici by ID
      tags:
      - Exercicis
...
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"